Restructure pager to show desktop names
In order to show desktop names, the pager needed to be restructured
quite a bit. Previously, there was a full representation that would use
the number of desktops as model, and a compact representation that would
only show one number box.
Since QML doesn't seem to have a way of directly accessing model data at
some index, the compact representation approach wouldn't work for
showing the desktop name.
Hence, we now have only one representation that handles both full and
compact and uses the actual pager model as model.